--- specs/transactions-full-stack.spec
@@ -1,5 +1,0 @@
-# Transactions
-
-## Add transactions
-
-* Add a transaction every "2" seconds for "10" seconds in mc_payments and mc_status
